requirements.txt
Flask==2.3.2
redis==4.6.0
flask-pymongo==2.3.0
